add template file
