+++ app/templates/404.html
@@ -1,0 +1,5 @@
+{% extends 'base.html'%}
+
+{% block content %}
+    <h1>WHOOOPS we can't find that page</h1>
+{% endblock %}
