+++ app/templates/base.html
@@ -1,0 +1,1 @@
+{% extends 'bootstrap/base.html'%}
