+++ app/templates/macros.html
@@ -1,0 +1,12 @@
+{% macro displayMovieList(movie_list) %}
+
+{% for movie in movie_list %}
+<div class="col-xs-12 col-sm-4 col-md-2 col-lg-2 movie-card">
+    <img src="{{ movie.poster }}" alt="{{ movie.title }} poster" class="img-responsive">
+    <li class="text-center">
+        <a href="/movie/{{movie.id}}">
+            {{ movie.title|truncate(10)}}</a>
+    </li>
+</div>
+{% endfor %}
+{% endmacro %}
